requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
httpx==0.27.0
anthropic==0.34.0
python-dotenv==1.0.1
pydantic==2.8.0
pandas==2.2.0
ta==0.11.0
aiohttp==3.10.0
